Switch to symbol-overlay from highlight-symbol and highlight combo
I have struggled to tune `hightlight-symbol` for a last few weeks. I've
encountered numerous issues, like automatically highlighted symbol not being
visible in current line (with `font-lock`'s background) due to conflict with
hl-line mode's overlay, jit-lock mode not flushing (required an advice
workaround), automatic highlighting clobbering `helpful-mode`
fontification (haven't got to the bottom of it), conflict in `highlihgt-symbol`
defun (both `highlight` and `highlight-symbol` defined it). It also seem to be
unmaintained - last version on MELPA is from 2016.

I've chosen to switch to `symbol-overlay`. Amongst others it defines its own
symbol highlighting. So I decided to use that instead of Exordium's custom one
that required `highlight`.

The last but not least `symbol-overlay` seems to be maintained and
`modus-themes` - that I use and intent to port to Exordium - has built in
support for it.