feat(company,git,forge): Add teams to completions in git/forge
Forge has added support for teams only
[recently](https://github.com/magit/forge/commit/7cdf978),
so lets take leverage of it. Only team handles (in a form of
`organisation/name`) are stored, hence no annotation with a team description in
the completion popup.

The backend suggest completion either when beginning of orgranisation matches
or when beginning of the team name matches.