feat(helm-projectile): add affixation function
This is done in an attempt to provide a consistent interface when switching to
a buffer, and is hidden behind `exordium-extra-help` flag.

This is done similarily to what Exordium does for `switch-to-buffer` and
`helpful` commands do under the hood. Both of these, when
`completions-detailed` is non-nil use `helm-completing-read-extra-metadata` to
select an affixation function.