feat(git,fill): Tune git-commit-ts-mode fill functions
Ensure that `fill-paragraph` (invoked with `M-q`) does the right thing, that is
to allow for a leading spaces in trailers and breaking changes.

Ensure that `auto-fill` (filling paragraphs as you type, when they extend the
`fill-column`) does the right thing that is to allow for a leading spaces in
trailers and breaking changes.

Use two spaces a offset which is an arbitrary value. Git specification requires
at least one space, but I think it looks well when the margin is slightly
larger.