refactor(company): Use external company-forge
I developed package [company-forge] starting with `exordium-company-assignees`
and extending its cababilities. The package, available on MELPA, provides
completions not only for assignees (users) but also for teams and for issues
and pull requests. It also provides icons for completions and uses
`quickhelp-string` and `doc-buffer` to provide more information about issues
and pull requests.

[company-forge] https://github.com/pkryger/company-forge.el