feat(elisp): Add interactive advice removal functions
In Emacs-30 `advice-remove` is an interactive function that asks user for the
symbol and the advice to remove. I found it to be very convenient when
experimenting with Emacs, hence porting.

The command `exordium-advice-remove-all` is similar, but it removes all advices
from a symbol.

Both commands use the same completing read handler as `helpful` use in
Exordium. This is to add extra permanent action that describes currently
selected candidate symbol.