feat(markdown): Don't breaking inline code in markdown when filling paragraphs
Although Github seems to be fine with it (to a point, see below), the inline
code is better to be kept in single line. When inline code is split into
multiple lines it doesn't render correctly in `markdown-mode`, i.e. it's not
fontified. Also Github joins such a split inline code when the text is
rendered, which can lead to misinterpretation after a paragraph has been
filled, e.g. should a line end with a first part of a split inline code, such a
line will be glued to the following line with a second part of the split inline
code. Besides, for longer code there's always a full code block available in
`markdown-mode`.