feat(markdown): Ignore invisible parts of links when filling paragraphs
When a paragraph is filled in `markdown-mode` the invisible parts of links
(e.g. the URL, the title) are ignored. This results in a paragraph that when
edited in Emacs looks like it has not been filled properly as lines that
contain links seem to long. However, this is closed to what Github will render,
thus reducing probability of strange looking posts.

A new option `exordium-ignore-links-predicate` has been added. It allows user
to control when to use the new behaviour. It's turned on in `markdown-mode`
(and in `gfm-mode`), as well as in `markdown-ts-mode`.  When set to nil, the
new behaviour will be disabled.