fix(helm): Ensure helm-buffer-max-length and helm-buffer-max-len-mode are honoured
Exordium uses Helm's completing read when switching buffers. A slightly
modified version of it. This is to ensure common look and feel when switching
buffers in other frame and other window. However, existing approach suffers
from truncating buffers names (i.e., falling back to normal value). This fix
uses approach from `helm-buffers-list--init` to ensure the
`helm-buffer-max-length` and `helm-buffer-max-len-mode` are obeyed in a case
when they are set to nil.